// src/features/dashboard-beasiswa/types/scholarship.types.ts
export interface IDataScholarship {
  _id: string;
  name: string;
  description: string;
  country: string;
  city: string;
  email: string;
  url_web: string;
  degrees: string;
  major: string;
  funding_type: string;
  open_date: string;
  close_date: string;
}

export interface IScholarship {
  name: string;
  description: string;
  country: string;
  city: string;
  email: string;
  url_web: string;
  degrees: string;
  major: string;
  funding_type: string;
  open_date: string;
  close_date: string;
}